requirements.txt
cffi==1.16.0
cryptography==42.0.1
pycparser==2.21
pycryptodomex==3.20.0
tqdm==4.66.1
